refactor: rename `Theme` to `Config`
