+++ src/tests/movement.rs
@@ -5,14 +5,14 @@ use crate::{
     component::task_list::TaskList,
     task::Task,
     tests::assert_task_eq,
-    theme::Theme,
+    theme::Config,
     utils::test::input_char,
 };
 
 #[test]
 fn test_rollover() {
     let mut app = App::new(
-        Theme::default(),
+        Config::default(),
         TaskStore {
             tasks: vec![
                 Task::from_string(String::from("meme")),
@@ -45,7 +45,7 @@ fn test_rollover() {
 #[test]
 fn test_shifting_tasks() {
     let mut app = App::new(
-        Theme::default(),
+        Config::default(),
         TaskStore {
             tasks: vec![
                 Task::from_string(String::from("meme")),
